gcp-streaming-pipeline/requirements.txt
# =============================================================================
# requirements.txt
# GCP Streaming Pipeline – Python dependencies
#
# Install with:
#   pip install -r requirements.txt
#
# For a reproducible environment, use a virtual environment first:
#   python -m venv .venv && source .venv/bin/activate
# =============================================================================

# ── Google Cloud ──────────────────────────────────────────────────────────────
# Pub/Sub client library (producer)
google-cloud-pubsub>=2.21.0

# BigQuery client library (ad-hoc queries, schema management)
google-cloud-bigquery>=3.20.0

# BigQuery Storage API (fast reads; optional but recommended)
google-cloud-bigquery-storage>=2.25.0

# ── Apache Beam ───────────────────────────────────────────────────────────────
# [gcp] extra pulls in all GCP-specific dependencies:
#   • google-cloud-dataflow-client
#   • grpcio, proto-plus, etc.
apache-beam[gcp]>=2.56.0

# ── Utilities ─────────────────────────────────────────────────────────────────
# Retry / back-off helpers (used in producer.py)
google-api-core>=2.19.0

# ── Testing ───────────────────────────────────────────────────────────────────
pytest>=8.2.0
pytest-cov>=5.0.0       # optional: coverage reports

# ── Development / linting (optional, comment out for CI) ─────────────────────
# black>=24.4.0
# pylint>=3.2.0
# mypy>=1.10.0

